+++ kuhub/views/follower_view.py
@@ -1,0 +1,44 @@
+from django.contrib.auth.decorators import login_required
+from django.http import JsonResponse
+from django.shortcuts import get_object_or_404, render
+from kuhub.models import Post, Profile, UserFollower
+from django.contrib.auth.models import User
+from django.views.decorators.http import require_POST
+from kuhub.views.firebase_view import navbar_setting_profile, separate_folder_firebase
+
+
+@login_required
+@require_POST
+def toggle_follow(request, user_id):
+    user_to_follow = User.objects.get(pk=user_id)
+    follower = request.user
+
+    is_following = UserFollower.objects.filter(user_followed=user_to_follow, follower=follower).exists()
+
+    if is_following:
+        # If already following, unfollow
+        UserFollower.objects.filter(user_followed=user_to_follow, follower=follower).delete()
+    else:
+        # If not following, follow
+        UserFollower.objects.create(user_followed=user_to_follow, follower=follower)
+
+    # Recalculate counts
+    followers_count = UserFollower.objects.filter(user_followed=user_to_follow).count()
+
+    return JsonResponse({'is_following': not is_following, 'followers_count': followers_count})
+
+
+@login_required
+def followers_page(request):
+    user = request.user
+    followers = UserFollower.objects.filter(user_followed=user)
+
+    return render(request, "kuhub/followers_page.html", context={'followers': followers})
+
+
+@login_required
+def following_page(request):
+    user = request.user
+    following = UserFollower.objects.filter(follower=user)
+
+    return render(request, "kuhub/following_page.html", context={'followings': following})
